Updated ResourceGuardProxy to ResourceGuard Mapping, re-tested
