+++ src/dataprotection/azext_dataprotection/aaz/latest/dataprotection/__cmd_group.py
@@ -13,6 +13,7 @@ from azure.cli.core.aaz import *
 
 @register_command_group(
     "dataprotection",
+    is_experimental=True,
 )
 class __CMDGroup(AAZCommandGroup):
     """Manage dataprotection.
