+++ src/dataprotection/azext_dataprotection/aaz/latest/dataprotection/backup_vault/__cmd_group.py
@@ -13,6 +13,7 @@ from azure.cli.core.aaz import *
 
 @register_command_group(
     "dataprotection backup-vault",
+    is_experimental=True,
 )
 class __CMDGroup(AAZCommandGroup):
     """Manage backup vault with dataprotection.
