+++ src/dataprotection/azext_dataprotection/aaz/latest/dataprotection/backup_vault/resource_guard_mapping/__init__.py
@@ -8,16 +8,9 @@
 # pylint: skip-file
 # flake8: noqa
 
-from azure.cli.core.aaz import *
-
-
-@register_command_group(
-    "dataprotection backup-vault resource-guard-proxy",
-)
-class __CMDGroup(AAZCommandGroup):
-    """Manage ResourceGuardProxies for a backup vault.
-    """
-    pass
-
-
-__all__ = ["__CMDGroup"]
+from .__cmd_group import *
+from ._create import *
+from ._delete import *
+from ._list import *
+from ._show import *
+from ._update import *
