+++ src/dataprotection/azext_dataprotection/aaz/latest/dataprotection/resource_guard/__init__.py
@@ -13,4 +13,5 @@ from ._create import *
 from ._delete import *
 from ._list import *
 from ._show import *
+from ._unlock import *
 from ._update import *
